Split core.treeView.Writer tests into separate files.
